More changes required for native compilation
These are Kafka binder changes required for native applications.
